% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntplib.R
\name{ntpHistogram}
\alias{ntpHistogram}
\title{Draw a simple histogram of some one variable in our log}
\usage{
ntpHistogram(log, start, end, output)
}
\arguments{
\item{start}{start time in strftime format}

\item{end}{end time in strftime format}

\item{output}{optional file in which to store the graph}

\item{dataframe}{returned by ntpLoopRead()}
}
\description{
Draw a simple histogram of some one variable in our log
}

