% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntplib.R
\name{ntpLoopRead}
\alias{ntpLoopRead}
\title{Read an NTPd loop file}
\usage{
ntpLoopRead(file)
}
\arguments{
\item{file}{A properly formatted NTP loop file generated by the daemon}
}
\value{
The raw data frame, merged timeseries (offset, delay, m->s, s->)
         offset time series, delay time series,
         master.to.slave time series, and slave.to.master time series
}
\description{
Read an NTPd loop file
}

