% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntplib.R
\name{ntpLoopStats}
\alias{ntpLoopStats}
\title{Textual output of the loop file's offset statistics}
\usage{
ntpLoopStats(log, boundary = 1000)
}
\arguments{
\item{boundary}{measure past which we are out of SLA in ns}

\item{dataframe}{returned by ntpLoopRead()}
}
\description{
Textual output of the loop file's offset statistics
}

