% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntplib.R
\name{ntpPeerRead}
\alias{ntpPeerRead}
\title{Read an NTPd peer file}
\usage{
ntpPeerRead(file)
}
\arguments{
\item{file}{A properly formatted NTP peer file generated by the daemon}
}
\value{
The raw data frame, merged timeseries (offset, drift, error)
}
\description{
Read an NTPd peer file
}

